fix(EDA-002): widen card misc state to Java int
Migrate CardInstance.misc and every production mutation path from i16 to i32, preserve Java wrapping arithmetic, and remove narrowing in the structured training snapshot restore path.

Java: decompiled/java-src/com/megacrit/cardcrawl/actions/common/ModifyDamageAction.java:21-29

Java: decompiled/java-src/com/megacrit/cardcrawl/cards/AbstractCard.java

Java: decompiled/java-src/com/megacrit/cardcrawl/cards/CardSave.java

Red: detached parent ./scripts/test_engine_rs.sh test --lib eda_002 -- --ignored --nocapture -> 0 passed, 1 failed (actual -32766, expected 32770).

Green: ./scripts/test_engine_rs.sh test --lib eda_002 -- --nocapture -> 2 passed; ./scripts/test_engine_rs.sh test --lib -> 2903 passed, 2 ignored; cargo check --all-targets clean.

Bench: full_turn_cycle 4.408us, clone_for_mcts 548ns, get_legal_actions 116.9ns.

+++ packages/engine-rs/src/training_contract.rs
@@ -949,7 +949,7 @@ fn build_card_token(
         target: format!("{:?}", def.target),
         cost_for_turn: card.cost.max(-1) as i32,
         base_cost: card.base_cost.max(-1) as i32,
-        misc: card.misc as i32,
+        misc: card.misc,
         upgraded: card.is_upgraded(),
         free_to_play: card.is_free(),
         retained: card.is_retained(),
@@ -969,7 +969,7 @@ fn build_card_snapshot(
         card_id: def.id.to_string(),
         cost_for_turn: card.cost.max(-1) as i32,
         base_cost: card.base_cost.max(-1) as i32,
-        misc: card.misc as i32,
+        misc: card.misc,
         upgraded: card.is_upgraded(),
         free_to_play: card.is_free(),
         retained: card.is_retained(),
@@ -987,7 +987,7 @@ fn restore_card_snapshots(
             let mut restored = registry.make_card(&card.card_id);
             restored.cost = card.cost_for_turn as i8;
             restored.base_cost = card.base_cost as i8;
-            restored.misc = card.misc as i16;
+            restored.misc = card.misc;
             restored.set_retained(card.retained);
             if card.upgraded {
                 restored.flags |= crate::combat_types::CardInstance::FLAG_UPGRADED;
